requirements.txt
flask
sqlalchemy
PyJWT
pytest
python-dotenv
